Moves class to separate file
